✨ 기능 수정 : WorkoutPerRoutine 구조 리팩토링
