+++ healthtart/src/main/java/com/dev5ops/healthtart/routine/controller/RoutineController.java
@@ -2,6 +2,12 @@ package com.dev5ops.healthtart.routine.controller;
 
 import com.dev5ops.healthtart.routine.domain.dto.RoutineDTO;
 import com.dev5ops.healthtart.routine.domain.vo.*;
+import com.dev5ops.healthtart.routine.domain.vo.request.RequestInsertRoutineVO;
+import com.dev5ops.healthtart.routine.domain.vo.request.RequestModifyRoutineVO;
+import com.dev5ops.healthtart.routine.domain.vo.response.ResponseDeleteRoutineVO;
+import com.dev5ops.healthtart.routine.domain.vo.response.ResponseFindRoutineVO;
+import com.dev5ops.healthtart.routine.domain.vo.response.ResponseInsertRoutineVO;
+import com.dev5ops.healthtart.routine.domain.vo.response.ResponseModifyRoutineVO;
 import com.dev5ops.healthtart.routine.service.RoutineServiceImpl;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
@@ -44,8 +50,8 @@ public class RoutineController {
 
     @PutMapping("/{routineCode}")
     @Operation(summary = "운동 루틴 수정")
-    public ResponseEntity<ResponseModifyRoutineVO> modifyRoutine( @PathVariable("routineCode") Long routineCode,
-            @RequestBody RequestModifyRoutineVO requestModify) {
+    public ResponseEntity<ResponseModifyRoutineVO> modifyRoutine(@PathVariable("routineCode") Long routineCode,
+                                                                 @RequestBody RequestModifyRoutineVO requestModify) {
         EditRoutineVO editRoutineVO = modelMapper.map(requestModify, EditRoutineVO.class);
         ResponseModifyRoutineVO response = routineService.modifyRoutine(routineCode, editRoutineVO);
         return ResponseEntity.ok(response);
