+++ healthtart/src/main/java/com/dev5ops/healthtart/routine/domain/vo/request/RequestDeleteRoutineVO.java
@@ -1,4 +1,4 @@
-package com.dev5ops.healthtart.routine.domain.vo;
+package com.dev5ops.healthtart.routine.domain.vo.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
