+++ healthtart/src/main/java/com/dev5ops/healthtart/routine/domain/vo/request/RequestModifyRoutineVO.java
@@ -1,19 +1,17 @@
-package com.dev5ops.healthtart.routine.domain.vo;
+package com.dev5ops.healthtart.routine.domain.vo.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-public class RequestInsertRoutineVO {
-    private Long routineCode;
+public class RequestModifyRoutineVO {
     private String title;
     private Integer time;
-    private String link;
-    private String recommendMusic;
-    private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
+
 }
