+++ healthtart/src/main/java/com/dev5ops/healthtart/workout_per_routine/controller/WorkoutPerRoutineController.java
@@ -1,7 +1,13 @@
 package com.dev5ops.healthtart.workout_per_routine.controller;
 
 import com.dev5ops.healthtart.workout_per_routine.domain.dto.WorkoutPerRoutineDTO;
-import com.dev5ops.healthtart.workout_per_routine.domain.vo.vo.*;
+import com.dev5ops.healthtart.workout_per_routine.domain.vo.*;
+import com.dev5ops.healthtart.workout_per_routine.domain.vo.request.RequestDeleteWorkoutPerRoutineVO;
+import com.dev5ops.healthtart.workout_per_routine.domain.vo.request.RequestInsertWorkoutPerRoutineVO;
+import com.dev5ops.healthtart.workout_per_routine.domain.vo.response.ResponseDeleteWorkoutPerRoutineVO;
+import com.dev5ops.healthtart.workout_per_routine.domain.vo.response.ResponseFindWorkoutPerRoutineVO;
+import com.dev5ops.healthtart.workout_per_routine.domain.vo.response.ResponseInsertWorkoutPerRoutineVO;
+import com.dev5ops.healthtart.workout_per_routine.domain.vo.response.ResponseModifyWorkoutPerRoutineVO;
 import com.dev5ops.healthtart.workout_per_routine.service.WorkoutPerRoutineServiceImpl;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
