+++ healthtart/src/main/java/com/dev5ops/healthtart/workout_per_routine/domain/vo/EditWorkoutPerRoutineVO.java
@@ -1,19 +1,18 @@
-package com.dev5ops.healthtart.workout_per_routine.domain.vo.vo;
+package com.dev5ops.healthtart.workout_per_routine.domain.vo;
+
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
-
-@AllArgsConstructor
 @NoArgsConstructor
+@AllArgsConstructor
 @Getter
-public class RequestModifyWorkoutPerRoutineVO {
+public class EditWorkoutPerRoutineVO {
     private int workoutOrder;
     private int weightSet;
     private int numberPerSet;
     private int weightPerSet;
     private int workoutTime;
-    private LocalDateTime updatedAt;
+
 }
