+++ healthtart/src/main/java/com/dev5ops/healthtart/workout_per_routine/domain/vo/request/RequestInsertWorkoutPerRoutineVO.java
@@ -1,5 +1,4 @@
-package com.dev5ops.healthtart.workout_per_routine.domain.vo.vo;
-
+package com.dev5ops.healthtart.workout_per_routine.domain.vo.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -10,7 +9,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-public class ResponseInsertWorkoutPerRoutineVO {
+public class RequestInsertWorkoutPerRoutineVO {
     private Long workoutPerRoutineCode;
     private int workoutOrder;
     private int weightSet;
