+++ healthtart/src/main/java/com/dev5ops/healthtart/workout_per_routine/domain/vo/request/RequestModifyWorkoutPerRoutineVO.java
@@ -1,4 +1,4 @@
-package com.dev5ops.healthtart.workout_per_routine.domain.vo.vo;
+package com.dev5ops.healthtart.workout_per_routine.domain.vo.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,17 +9,11 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-public class RequestInsertWorkoutPerRoutineVO {
-    private Long workoutPerRoutineCode;
+public class RequestModifyWorkoutPerRoutineVO {
     private int workoutOrder;
     private int weightSet;
     private int numberPerSet;
     private int weightPerSet;
     private int workoutTime;
-    private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    private Long recordCode;
-    private Long exerciseEquipmentCode;
-
-
 }
