+++ healthtart/src/main/java/com/dev5ops/healthtart/workout_per_routine/domain/vo/response/ResponseDeleteWorkoutPerRoutineVO.java
@@ -1,4 +1,4 @@
-package com.dev5ops.healthtart.workout_per_routine.domain.vo.vo;
+package com.dev5ops.healthtart.workout_per_routine.domain.vo.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,12 +9,17 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-public class ResponseModifyWorkoutPerRoutineVO {
+public class ResponseDeleteWorkoutPerRoutineVO {
+    private Long workoutPerRoutineCode;
     private int workoutOrder;
     private int weightSet;
     private int numberPerSet;
     private int weightPerSet;
     private int workoutTime;
+    private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
+    private Long recordCode;
+    private Long exerciseEquipmentCode;
+
 
 }
