+++ healthtart/src/main/java/com/dev5ops/healthtart/workout_per_routine/domain/vo/response/ResponseFindWorkoutPerRoutineVO.java
@@ -1,4 +1,4 @@
-package com.dev5ops.healthtart.workout_per_routine.domain.vo.vo;
+package com.dev5ops.healthtart.workout_per_routine.domain.vo.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
