+++ healthtart/src/main/java/com/dev5ops/healthtart/workout_per_routine/domain/vo/response/ResponseInsertWorkoutPerRoutineVO.java
@@ -1,4 +1,5 @@
-package com.dev5ops.healthtart.workout_per_routine.domain.vo.vo;
+package com.dev5ops.healthtart.workout_per_routine.domain.vo.response;
+
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,7 +10,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-public class ResponseDeleteWorkoutPerRoutineVO {
+public class ResponseInsertWorkoutPerRoutineVO {
     private Long workoutPerRoutineCode;
     private int workoutOrder;
     private int weightSet;
