+++ healthtart/src/main/java/com/dev5ops/healthtart/workout_per_routine/domain/vo/response/ResponseModifyWorkoutPerRoutineVO.java
@@ -1,18 +1,20 @@
-package com.dev5ops.healthtart.workout_per_routine.domain.vo.vo;
-
+package com.dev5ops.healthtart.workout_per_routine.domain.vo.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@NoArgsConstructor
+import java.time.LocalDateTime;
+
 @AllArgsConstructor
+@NoArgsConstructor
 @Getter
-public class EditWorkoutPerRoutineVO {
+public class ResponseModifyWorkoutPerRoutineVO {
     private int workoutOrder;
     private int weightSet;
     private int numberPerSet;
     private int weightPerSet;
     private int workoutTime;
+    private LocalDateTime updatedAt;
 
 }
