+++ healthtart/src/test/java/com/dev5ops/healthtart/routine/service/RoutineServiceImplTests.java
@@ -5,6 +5,10 @@ import com.dev5ops.healthtart.common.exception.StatusEnum;
 import com.dev5ops.healthtart.routine.domain.dto.RoutineDTO;
 import com.dev5ops.healthtart.routine.domain.entity.Routine;
 import com.dev5ops.healthtart.routine.domain.vo.*;
+import com.dev5ops.healthtart.routine.domain.vo.response.ResponseDeleteRoutineVO;
+import com.dev5ops.healthtart.routine.domain.vo.response.ResponseFindRoutineVO;
+import com.dev5ops.healthtart.routine.domain.vo.response.ResponseInsertRoutineVO;
+import com.dev5ops.healthtart.routine.domain.vo.response.ResponseModifyRoutineVO;
 import com.dev5ops.healthtart.routine.repository.RoutineRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
